MNT: wrap generator implementation inside reference implementation
