+++ skxray/core/accumulators/correlation.py
@@ -37,7 +37,6 @@ from __future__ import absolute_import, division, print_function
 
 from skxray.core.utils import multi_tau_lags
 from skxray.core.roi import extract_label_indices
-from skxray.core.correlation import correlation as corr
 from collections import namedtuple
 import numpy as np
 
@@ -105,15 +104,72 @@ class InternalCorrelationState:
         self.processed = 0
 
 
-def _process(_state, num_bufs, num_pixels, level, buf_no):
-    corr._process(_state.buf, _state.G, _state.past_intensity,
-                  _state.future_intensity, _state.label_mask,
-                  num_bufs, num_pixels, _state.img_per_level,
-                  level, buf_no)
+def _process(buf, G, past_intensity_norm, future_intensity_norm,
+             label_mask, num_bufs, num_pixels, img_per_level, level, buf_no):
+    """Internal helper function.
+
+    This helper function calculates G, past_intensity_norm and
+    future_intensity_norm at each level, symmetric normalization is used.
+
+    .. warning :: This modifies inputs in place.
+
+    Parameters
+    ----------
+    buf : array
+        image data array to use for correlation
+    G : array
+        matrix of auto-correlation function without normalizations
+    past_intensity_norm : array
+        matrix of past intensity normalizations
+    future_intensity_norm : array
+        matrix of future intensity normalizations
+    label_mask : array
+        labeled array where all nonzero values are ROIs
+    num_bufs : int, even
+        number of buffers(channels)
+    num_pixels : array
+        number of pixels in certain roi's
+        roi's, dimensions are : [number of roi's]X1
+    img_per_level : array
+        to track how many images processed in each level
+    level : int
+        the current multi-tau level
+    buf_no : int
+        the current buffer number
+
+    Notes
+    -----
+    :math ::
+        G   = <I(\tau)I(\tau + delay)>
+    :math ::
+        past_intensity_norm = <I(\tau)>
+    :math ::
+        future_intensity_norm = <I(\tau + delay)>
+    """
+    img_per_level[level] += 1
+    # in multi-tau correlation, the subsequent levels have half as many
+    # buffers as the first
+    i_min = num_bufs // 2 if level else 0
+
+    for i in range(i_min, min(img_per_level[level], num_bufs)):
+        # compute the index into the autocorrelation matrix
+        t_index = level * num_bufs / 2 + i
+
+        delay_no = (buf_no - i) % num_bufs
+        # get the images for correlating
+        past_img = buf[level, delay_no]
+        future_img = buf[level, buf_no]
+        for w, arr in zip([past_img*future_img, past_img, future_img],
+                          [G, past_intensity_norm, future_intensity_norm]):
+            binned = np.bincount(label_mask, weights=w)
+            # pdb.set_trace()
+            arr[t_index] += ((binned / num_pixels - arr[t_index]) /
+                             (img_per_level[level] - i))
+    return None  # modifies arguments in place!
 
 
-def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
-                     _state=None):
+def lazy_multi_tau(image_iterable, num_levels, num_bufs, labels,
+                   processing_func=_process, _state=None):
     """Generator implementation of 1-time multi-tau correlation
 
     Parameters
@@ -134,6 +190,9 @@ def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
         example, if you have four ROIs, they must be labeled 1, 2, 3,
         4. Background is labeled as 0
     images : iterable of 2D arrays
+    _processing_func : function, optional
+        The processing function for the internals of the lazy correlator.
+        Defaults to skxray.core.accumulators.correlation._process
     _state : namedtuple, optional
         _state is a bucket for all of the internal state of the generator.
         It is part of the `results` object that is yielded from this
@@ -147,6 +206,32 @@ def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
         - the times at which the correlation was computed, `lag_steps`
         - and all of the internal state, `final_state`, which is a
           `correlation_state` namedtuple
+
+    Notes
+    -----
+
+    The normalized intensity-intensity time-autocorrelation function
+    is defined as
+
+    :math ::
+        g_2(q, t') = \frac{<I(q, t)I(q, t + t')> }{<I(q, t)>^2}
+
+    ; t' > 0
+
+    Here, I(q, t) refers to the scattering strength at the momentum
+    transfer vector q in reciprocal space at time t, and the brackets
+    <...> refer to averages over time t. The quantity t' denotes the
+    delay time
+
+    This implementation is based on published work. [1]_
+
+    References
+    ----------
+
+    .. [1] D. Lumma, L. B. Lurio, S. G. J. Mochrie and M. Sutton,
+        "Area detector based photon correlation in the regime of
+        short data batches: Data reduction for dynamic x-ray
+        scattering," Rev. Sci. Instrum., vol 70, p 3274-3289, 2000.
     """
     if _state is None:
         _state = InternalCorrelationState(num_levels, num_bufs, labels)
@@ -173,7 +258,9 @@ def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
         # (undownsampled) frames. This modifies G,
         # past_intensity, future_intensity,
         # and img_per_level in place!
-        _process(s, num_bufs, num_pixels, level, buf_no)
+        processing_func(s.buf, s.G, s.past_intensity, s.future_intensity,
+                        s.label_mask, num_bufs, num_pixels, s.img_per_level,
+                        level, buf_no)
 
         # check whether the number of levels is one, otherwise
         # continue processing the next level
@@ -203,7 +290,9 @@ def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
                 # than one. This is modifying things in place. See comment
                 # on previous call above.
                 buf_no = s.cur[level] - 1
-                _process(s, num_bufs, num_pixels, level, buf_no)
+                processing_func(s.buf, s.G, s.past_intensity,
+                                s.future_intensity, s.label_mask, num_bufs,
+                                num_pixels, s.img_per_level, level, buf_no)
                 level += 1
 
                 # Checking whether there is next level for processing
@@ -221,6 +310,6 @@ def lazy_correlation(image_iterable, num_levels, num_bufs, labels,
         g2 = (s.G[:g_max] /
               (s.past_intensity[:g_max] *
                s.future_intensity[:g_max]))
-        
+
         s.processed += 1
         yield results(g2, lag_steps[:g_max], s)
