+++ skxray/core/correlation/correlation.py
@@ -46,285 +46,24 @@ This module is for functions specific to time correlation
 """
 from __future__ import absolute_import, division, print_function
 import logging
-import time
 
 import numpy as np
 
-from .. import utils as core
-from .. import roi
+from ..accumulators.correlation import lazy_multi_tau, _process
 
 logger = logging.getLogger(__name__)
 
 
-def _process(buf, G, past_intensity_norm, future_intensity_norm,
-             label_mask, num_bufs, num_pixels, img_per_level, level, buf_no):
-    """Internal helper function.
-
-    This helper function calculates G, past_intensity_norm and
-    future_intensity_norm at each level, symmetric normalization is used.
-
-    .. warning :: This modifies inputs in place.
-
-    Parameters
-    ----------
-    buf : array
-        image data array to use for correlation
-    G : array
-        matrix of auto-correlation function without normalizations
-    past_intensity_norm : array
-        matrix of past intensity normalizations
-    future_intensity_norm : array
-        matrix of future intensity normalizations
-    label_mask : array
-        labeled array where all nonzero values are ROIs
-    num_bufs : int, even
-        number of buffers(channels)
-    num_pixels : array
-        number of pixels in certain roi's
-        roi's, dimensions are : [number of roi's]X1
-    img_per_level : array
-        to track how many images processed in each level
-    level : int
-        the current multi-tau level
-    buf_no : int
-        the current buffer number
-
-    Notes
-    -----
-    :math ::
-        G   = <I(\tau)I(\tau + delay)>
-    :math ::
-        past_intensity_norm = <I(\tau)>
-    :math ::
-        future_intensity_norm = <I(\tau + delay)>
-    """
-    img_per_level[level] += 1
-    # in multi-tau correlation, the subsequent levels have half as many
-    # buffers as the first
-    i_min = num_bufs // 2 if level else 0
-
-    for i in range(i_min, min(img_per_level[level], num_bufs)):
-        # compute the index into the autocorrelation matrix
-        t_index = level * num_bufs / 2 + i
-
-        delay_no = (buf_no - i) % num_bufs
-        # get the images for correlating
-        past_img = buf[level, delay_no]
-        future_img = buf[level, buf_no]
-        for w, arr in zip([past_img*future_img, past_img, future_img],
-                          [G, past_intensity_norm, future_intensity_norm]):
-            binned = np.bincount(label_mask, weights=w)
-            # pdb.set_trace()
-            arr[t_index] += ((binned / num_pixels - arr[t_index]) /
-                             (img_per_level[level] - i))
-    return None  # modifies arguments in place!
-
-
 def multi_tau_auto_corr(num_levels, num_bufs, labels, images,
                         processing_func=_process):
+    """Wraps generator implementation of multi-tau
+
+    See skxray.core.accumulators.correlation.lazy_correlation
     """
-    This function computes one-time correlations.
-
-    It uses a scheme to achieve long-time correlations inexpensively
-    by downsampling the data, iteratively combining successive frames.
-
-    The longest lag time computed is num_levels * num_bufs.
-
-    Parameters
-    ----------
-    num_levels : int
-        how many generations of downsampling to perform, i.e., the depth of
-        the binomial tree of averaged frames
-
-    num_bufs : int, must be even
-        maximum lag step to compute in each generation of downsampling
-
-    labels : array
-        Labeled array of the same shape as the image stack.
-        Each ROI is represented by sequential integers starting at one.  For
-        example, if you have four ROIs, they must be labeled 1, 2, 3,
-        4. Background is labeled as 0
-
-    images : iterable of 2D arrays
-        dimensions are: (rr, cc)
-
-    processing_func : function
-        The function used in the inner loop of multi_tau_auto_corr.
-        Defaults to the reference python implementation (note that it is
-        optimized for clarity, not speed) in
-        skxray.core.correlation.correlation:_process
-
-    Returns
-    -------
-    g2 : array
-        Matrix of normalized intensity-intensity autocorrelation.
-        Shape is (num_levels, np.unique(labels)))
-
-    lag_steps : array
-        Delay or lag steps for the multiple tau analysis.
-        Shape is num_levels
-
-    Notes
-    -----
-
-    The normalized intensity-intensity time-autocorrelation function
-    is defined as
-
-    :math ::
-        g_2(q, t') = \frac{<I(q, t)I(q, t + t')> }{<I(q, t)>^2}
-
-    ; t' > 0
-
-    Here, I(q, t) refers to the scattering strength at the momentum
-    transfer vector q in reciprocal space at time t, and the brackets
-    <...> refer to averages over time t. The quantity t' denotes the
-    delay time
-
-    This implementation is based on published work. [1]_
-
-    References
-    ----------
-
-    .. [1] D. Lumma, L. B. Lurio, S. G. J. Mochrie and M. Sutton,
-        "Area detector based photon correlation in the regime of
-        short data batches: Data reduction for dynamic x-ray
-        scattering," Rev. Sci. Instrum., vol 70, p 3274-3289, 2000.
-    """
-    # In order to calculate correlations for `num_bufs`, images must be
-    # kept for up to the maximum lag step. These are stored in the array
-    # buffer. This algorithm only keeps number of buffers and delays but
-    # several levels of delays number of levels are kept in buf. Each
-    # level has twice the delay times of the next lower one. To save
-    # needless copying, of cyclic storage of images in buf is used.
-
-    if num_bufs % 2 != 0:
-        raise ValueError("num_bufs must be even. You provided %s" % num_bufs)
-
-    if hasattr(images, 'frame_shape'):
-        # Give a user-friendly error if we can detect the shape from pims.
-        if labels.shape != images.frame_shape:
-            raise ValueError("Shape of the image stack should be equal to"
-                             " shape of the labels array")
-
-    # get the pixels in each label
-    label_mask, pixel_list = roi.extract_label_indices(labels)
-
-    # Convert from num_levels, num_bufs to lag frames.
-    tot_channels, lag_steps = core.multi_tau_lags(num_levels, num_bufs)
-
-    # number of pixels per ROI
-    # TODO: Verify that this logic is ok.  It means that the ROIs **must** be
-    # integers that start with 1 and are sequential. Best option is to map the
-    # indices onto a sequential list of integers starting at 1
-    labels = np.unique(label_mask)
-    for n, label in enumerate(labels):
-        label_mask[label_mask == label] = n
-    num_pixels = np.bincount(label_mask)
-    # num_pixels = num_pixels[1:]
-    # pdb.set_trace()
-
-    try:
-        num_images = len(images)
-    except TypeError:
-        num_images = -1
-
-    # G holds the un normalized auto- correlation result. We
-    # accumulate computations into G as the algorithm proceeds.
-    G = np.zeros(((num_levels + 1) * num_bufs / 2, len(labels)),
-                 dtype=np.float64)
-
-    # matrix of past intensity normalizations
-    past_intensity_norm = np.zeros_like(G)
-
-    # matrix of future intensity normalizations
-    future_intensity_norm = np.zeros_like(G)
-
-    # Ring buffer, a buffer with periodic boundary conditions.
-    # Images must be keep for up to maximum delay in buf.
-    buf = np.zeros((num_levels, num_bufs, np.sum(num_pixels)),
-                   dtype=np.float64)
-
-    # to track processing each level
-    track_level = np.zeros(num_levels, dtype=bool)
-
-    # to increment buffer
-    cur = np.ones(num_levels, dtype=np.int64)
-
-    # to track how many images processed in each level
-    img_per_level = np.zeros(num_levels, dtype=np.int64)
-
-    start_time = time.time()  # log computation time to INFO
-
-    for img_num, img in enumerate(images):
-
-        cur[0] = (1 + cur[0]) % num_bufs  # increment buffer
-
-        # Put the ROI pixels into the ring buffer.
-        buf[0, cur[0] - 1] = np.ravel(img)[pixel_list]
-
-        # Compute the correlations between the first level
-        # (undownsampled) frames. This modifies G,
-        # past_intensity_norm, future_intensity_norm,
-        # and img_per_level in place!
-        buf_no = cur[0] - 1
-        processing_func(buf, G, past_intensity_norm,
-                        future_intensity_norm, label_mask,
-                        num_bufs, num_pixels, img_per_level,
-                        0, buf_no=buf_no)
-
-        # check whether the number of levels is one, otherwise
-        # continue processing the next level
-        processing = num_levels > 1
-
-        # Compute the correlations for all higher levels.
-        level = 1
-        prev = None
-        while processing:
-            if not track_level[level]:
-                track_level[level] = True
-                processing = False
-            else:
-                prev = 1 + (cur[level - 1] - 2) % num_bufs
-                cur[level] = 1 + cur[level] % num_bufs
-
-                buf[level, cur[level] - 1] = (
-                    (buf[level - 1, prev - 1] +
-                     buf[level - 1, cur[level - 1] - 1]) / 2
-                )
-
-                # make the track_level zero once that level is processed
-                track_level[level] = False
-
-                # call processing_func for each multi-tau level greater
-                # than one. This is modifying things in place. See comment
-                # on previous call above.
-                buf_no = cur[level] - 1
-                processing_func(buf, G, past_intensity_norm,
-                                future_intensity_norm, label_mask,
-                                num_bufs, num_pixels, img_per_level,
-                                level, buf_no=buf_no)
-                level += 1
-
-                # Checking whether there is next level for processing
-                processing = level < num_levels
-
-    # ending time for the process
-    end_time = time.time()
-
-    logger.info("Processing time for %s images took %s seconds.",
-                img_num, (end_time - start_time))
-
-    # the normalization factor
-    if len(np.where(past_intensity_norm == 0)[0]) != 0:
-        g_max = np.where(past_intensity_norm == 0)[0][0]
-    else:
-        g_max = past_intensity_norm.shape[0]
-
-    # g2 is normalized G
-    g2 = (G[:g_max] / (past_intensity_norm[:g_max] *
-                       future_intensity_norm[:g_max]))
-
-    return g2, lag_steps[:g_max]
+    gen = lazy_multi_tau(images, num_levels, num_bufs, labels,
+                         processing_func)
+    final_res = list(gen)[-1]
+    return final_res.g2, final_res.lag_steps
 
 
 def auto_corr_scat_factor(lags, beta, relaxation_rate, baseline=1):
